simplify metric by removing visittracking as no longer needed
