+++ src/app/components/OptimizelyPageMetrics/PageViewTracking/index.tsx
@@ -1,9 +1,74 @@
 import { useState, useContext, useEffect } from 'react';
 import { OptimizelyContext } from '@optimizely/react-sdk';
-import { getVisitTrackingState, recordVisitActivity } from '../VisitTracking';
 
 const PAGE_VIEW_EVENT_NAME = 'page-views';
 const VISIT_EVENT_NAME = 'visit';
+
+const VISIT_STORAGE_KEY = 'last_visit_ts';
+// 60 minute inactivity window defines a new visit for this spike
+// activity is defined as a page view, so the window rolls forward on each page view
+const VISIT_TIMEOUT_MS = 60 * 60 * 1000;
+
+// guard local storage access for privacy modes
+const getStorage = (): Storage | null => {
+  if (typeof window === 'undefined') {
+    return null;
+  }
+
+  try {
+    return window.localStorage;
+  } catch {
+    return null;
+  }
+};
+
+// parse a stored value into a usable timestamp
+const parseTimestamp = (value: string | null) => {
+  if (!value) {
+    return null;
+  }
+
+  const timestamp = Number(value);
+  return Number.isFinite(timestamp) && timestamp > 0 ? timestamp : null;
+};
+
+// read the last recorded activity from storage
+const readLastActivity = (storage: Storage | null) => {
+  if (!storage) {
+    return null;
+  }
+
+  try {
+    return parseTimestamp(storage.getItem(VISIT_STORAGE_KEY));
+  } catch {
+    return null;
+  }
+};
+
+// compute whether the current activity starts a new visit window
+// this is based on the last recorded activity, not the first, so the window rolls forward
+const getVisitTrackingState = (now: number) => {
+  const storage = getStorage();
+  const lastActivity = readLastActivity(storage);
+  const isNewVisit =
+    !lastActivity || now < lastActivity || now - lastActivity > VISIT_TIMEOUT_MS;
+
+  return { storage, isNewVisit };
+};
+
+// update the stored activity so ongoing browsing stays in one visit
+// this should be called on every page view to keep the visit alive
+const recordVisitActivity = (storage: Storage | null, now: number) => {
+  if (!storage) {
+    return;
+  }
+
+  try {
+    storage.setItem(VISIT_STORAGE_KEY, String(now));
+  } catch {
+    // swallow storage errors to avoid breaking tracking
+  }
+};
 
 // optional flag to control visit tracking alongside page views
 type Props = {
